My First Project.v1i.yolov8/requirements.txt
# Requirements for Ring Detection System - Complete Setup
# Install with: pip install -r requirements.txt

# Core Detection
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Configuration
pyyaml>=6.0

# Cloud & Database
firebase-admin>=6.0.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.0.0
pandas>=2.0.0

# IP Camera (Optional)
paho-mqtt>=1.6.1

# Utilities
requests>=2.28.0
python-dateutil>=2.8.2
